% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BODEindex.R
\name{bode}
\alias{bode}
\title{The BODE Index in COPD}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/14999112}
}
\usage{
bode(FEV1, mMRC, BMI, walk)
}
\arguments{
\item{FEV1}{A number for forced expiratory volume in one second (FEV1) expressed as percentage}

\item{mMRC}{A number for the dyspnea scale of modified Medical Research Council (mMRC)}

\item{BMI}{A number for Body Mass Index}

\item{walk}{A number for distance walked in 6 minutes (in meter)}
}
\description{
The BODE Index in COPD
}
\examples{
bode (FEV1 = 40, mMRC = 3, BMI = 22, walk = 100)
bode (FEV1 = 60, mMRC = 1, BMI = 19, walk = 200)

}
